final commit

+++ bras/app/components/nifty.py
@@ -1,30 +1,16 @@
 """Utilities dedicated to files (.nii)"""
-from typing import List
-from enum import Enum
+import enum
 import logging
-from pathlib import Path
-from fastapi import UploadFile
-import nibabel as nib
-import numpy as np
-import cv2
 import shutil
 import aiofiles
-
-
-class NiftyProcessor:
-    """read .nii files and convert into stack of uint8 images"""
-
-    def normalize_into_uint8(self, nib_3d: np.array):
-        """normaize pixel values into range 0-255 and return as type np.uint8"""
-        logging.info("normalizing nifty data to uint8")
-        normalized_images: np.array = cv2.normalize(
-            nib_3d, None, alpha=0, beta=255, norm_type=cv2.NORM_MINMAX, dtype=cv2.CV_32F)
-        return normalized_images.astype(np.uint8)
-
-    def reconstruct_stack(self, nib_3d: np.array) -> List:
-        """convert (W, H, Stack) into (Stack, H, W) and return list type"""
-        logging.info("re-construct nifty data")
-        return np.moveaxis(nib_3d, -1, 0).tolist()
+import numpy as np
+from pathlib import Path
+from fastapi import UploadFile
+from bras.nn.vino import VinoBraTs, normalize
+from bras.utils.image import MRI
+from monai.inferers import sliding_window_inference
+import nibabel as nib
+import torch
 
 
 class NiftyFileManager:
@@ -51,8 +37,9 @@ class NiftyFileManager:
         return file_path
 
     def remove_temporary_directory(self):
-        logging.info("clear temporary directory")
-        shutil.rmtree(self.temp_directory)
+        if Path(self.temp_directory).exists():        
+            logging.info("clear temporary directory")
+            shutil.rmtree(self.temp_directory)
 
 
 class NiftyFileReader:
@@ -61,23 +48,75 @@ class NiftyFileReader:
         logging.info("read nifty data from {}".format(path))
         return nib.load(path).get_fdata()
 
+    def return_stack_and_normalize(self, t1, t1ce, t2, flair):
+        def T(x): return normalize(
+            image=x,
+            mask=x > 0,
+            full_intensities_range=True
+        )
 
-class NiftyViewManager(
-    NiftyProcessor,
-    NiftyFileManager,
-    NiftyFileReader
-):
-    def __init__(self) -> None:
-        super().__init__()
+        return np.stack([
+            T(t1), T(t1ce), T(t2), T(flair)
+        ])
 
-    async def get_view(self, file: UploadFile):
-        logging.info("start processing new nifty file upload..")
-        # read file into numpy array
-        nib_file_path = await self.save_and_return_path(file)
-        nib_3d = self.read_nib(nib_file_path)
-        # normalize to later be manipulated into image
-        normalized_nib = self.normalize_into_uint8(nib_3d)
-        view = self.reconstruct_stack(normalized_nib)
+    def merge_segmentations(self, segmentations: np.array):
+        mask = np.zeros(segmentations.shape[1:])
+        for idx, channel in enumerate(segmentations, start=1):
+            mask[np.where(channel > 0.62)] = idx
 
+        return mask
+
+class SegmentationController(NiftyFileManager, NiftyFileReader, VinoBraTs):
+    def __init__(self, path_to_onnx) -> None:
+        super().__init__(
+            path_to_onnx=path_to_onnx
+        )
+
+    @staticmethod
+    def sigmoid_function(z):
+        """ this function implements the sigmoid function, and 
+        expects a numpy array as argument """
+        sigmoid = 1.0/(1.0 + np.exp(-z))
+        return sigmoid 
+
+    async def process_segmentation(self, t1, t1ce, t2, flair):
         self.remove_temporary_directory()
-        return view
+
+        async def process_nib(x):
+            path = await self.save_and_return_path(x)
+            return self.read_nib(path=path)
+
+        t1_weights = await process_nib(t1)
+        t1ce_weights = await process_nib(t1ce)
+        t2_weights = await process_nib(t2)
+        flair_weights = await process_nib(flair)
+
+        logging.info("stacking all input weights..")
+        model_inputs = self.return_stack_and_normalize(
+            t1=t1_weights,
+            t1ce=t1ce_weights,
+            t2=t2_weights,
+            flair=flair_weights
+        )
+
+
+        logging.info("feed inputs into segmentation model")
+        T = lambda x: torch.Tensor(x)
+        model_output = sliding_window_inference(
+            inputs=T(model_inputs.reshape((1, *model_inputs.shape))),
+            roi_size=(128, 128, 128),
+            sw_batch_size=1,
+            predictor= lambda x: T(self.inference(x.numpy())),
+            overlap=0.5
+        )
+
+        model_output = torch.sigmoid(model_output)
+        segmentation = self.merge_segmentations(model_output[0])
+        ni_img = nib.Nifti1Image(segmentation, affine=np.eye(4))
+        
+        save_path = Path(self.temp_directory) / "result.nii.gz"
+        nib.save(ni_img, save_path)
+        return save_path
+        
+
+
